# src/main/resources/api.yaml
openapi: "3.0.3"
info:
  title: Biblioteca API
  description: API para gerenciar organizações
  version: 1.0.0
...
paths:
  /auth/revoke:
    post:
      summary: Revoke token
      description: Revoke the authorization token.
      tags:
        - Auth
      operationId: revokeToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: "eyJhbGciOi"
      responses:
        '200':
          description: Token revoked successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token revoked successfully."
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid request."
  /auth/verify-authorization:
    get:
      summary: Verify authorization endpoint
      description: Verifica a autorização do token.
      tags:
        - Auth
      operationId: verifyAuthorization
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                type: boolean
                example: true
        '400':
          description: Token inválido ou ausente
          content:
            application/json:
              schema:
                type: boolean
                example: false
  /auth/login:
    post:
      summary: Login endpoint
      description: Endpoint para autenticação de usuários.
      tags:
        - Auth
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Credenciais inválidas."
  /auth/register:
    post:
      summary: Register endpoint
      description: Endpoint para registrar novos usuários.
      tags:
        - Auth
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
      responses:
        '200':
          description: Registro bem-sucedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Usuário já cadastrado."
  /chat/generate:
    post:
      summary:  Generate endpoint
      description: Endpoint para gerar resposta baseada no modelo e prompt fornecidos.
      tags:
        - chat
      operationId: generate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequestDTO'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponseDTO'
...
  /cidade:
    get:
      summary: Busca Cidades
      security:
        - bearerAuth: []
      tags:
        - Cidade
      operationId: findCidade
      parameters:
        - name: nome
          in: path
          required: false
          schema:
            type: string
        - name: estado_id
          in: path
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CidadeModel'
  /cidade/{id}:
    get:
      summary: Detalha uma Cidade
      security:
        - bearerAuth: []
      tags:
        - Cidade
      operationId: detailCidade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CidadeModel'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # Especifica o formato do token, pode ser omitido
  schemas:
    ChatRequestDTO:
      type: object
      properties:
        model:
          type: string
        prompt:
          type: string
    LoginRequestDTO:
      type: object
      properties:
        email:
          type: string
          format: email
        senha:
          type: string
...
    ChatResponseDTO:
        type: object
        properties:
            model:
              type: string
            createdAt:
              type: string
            response:
              type: string
            done:
                type: boolean
    LoginResponseDTO:
      type: object
      properties:
        nome:
          type: string
        authorization:
          type: string
...
    RegisterRequestDTO:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        senha:
          type: string
        dataNascimento:
          type: string
          format: date
        enderecos:
          type: array
          items:
            $ref: '#/components/schemas/EnderecoModel'
        perfis:
          type: array
          items:
            $ref: '#/components/schemas/PerfilModel'
...
    EnderecoModel:
      type: object
      properties:
        id:
          type: integer
        logradouro:
          type: string
        numero:
          type: integer
        complemento:
            type: string
        bairro:
          type: string
        cidade:
          type: string
        uf:
          $ref: '#/components/schemas/UnidadeFederativaModel'
        cep:
          type: string
...
    PerfilModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
    UnidadeFederativaModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sigla:
            type: string
    CidadeModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        estado_id:
          $ref: '#/components/schemas/EstadoModel'
    EstadoModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        pais_id:
          $ref: '#/components/schemas/PaisModel'
    PaisModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string